Refactor rails routes
