+++ lib/routing_constraints.rb
@@ -1,0 +1,1 @@
+require 'routing_constraints/format'
