+++ lib/routing_constraints/format.rb
@@ -1,0 +1,13 @@
+module RoutingConstraints
+  class Format
+    attr_accessor :mime_type
+
+    def initialize(format)
+      @mime_type = Mime::Type.lookup_by_extension(format)
+    end
+
+    def matches?(request)
+      request.format == mime_type
+    end
+  end
+end
